backend/pyproject.toml
[project]
name = "live-ai-video-tutor-backend"
version = "0.1.0"
description = "FastAPI backend for the live AI video tutor"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "httpx>=0.28.0",
  "python-dotenv>=1.1.0",
  "pydantic>=2.11.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=6.0.0",
]

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=100"
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = [
  "app",
  "pyproject.toml",
]
